requirements.txt
streamlit
watchdog
anthropic
pillow
starlette
uvicorn
httpx
fastapi
requests
pillow
fastmcp
dotenv
jsonref
ollama
openai
openai-agents
azure-ai-projects
azure-identity
mysql-connector-python
cryptography
snowflake  
snowflake-connector-python
snowflake-snowpark-python
snowflake[ml]
minio
pyinstaller
pandas
xlsxwriter
openpyxl
xlrd
python-docx
bs4
jinja2
langchain
langchain_community
langchain_core
black
build
PyMuPDF
tox
pypdf
